+++ include/rt/utils.hpp
@@ -39,7 +39,7 @@ using namespace Eigen;
 namespace rt
 {
 	/// A tolerance threshold.
-	const float EPSILON = 1e-4f;
+	const double EPSILON = 1e-4f;
 
 	/* A few useful constants */
 	#undef M_PI
@@ -52,40 +52,40 @@ namespace rt
 	#define INV_SQRT_TWO 0.70710678118654752440f
 
 	/// Values between -epsilon and epsilon are considered to be zero.
-	inline bool is_zero(const float x)
+	inline bool is_zero(const double x)
 	{
 		if ((x < EPSILON) && (x > -EPSILON)) return true; else return false;
 	}
 
 	/// Returns true if the arguments are equal within epsilon tolerance.
-	inline bool is_equal(const float x, const float y)
+	inline bool is_equal(const double x, const double y)
 	{
 		return is_zero(x-y);
 	}
 
 	/// Returns true if the vectors components are equal between epsilon tolerance
-	inline bool is_equal(const Vector3f& v1, const Vector3f& v2) 
+	inline bool is_equal(const Vector3d& v1, const Vector3d& v2) 
 	{
-		float v1_x = v1.data()[0];
-		float v1_y = v1.data()[1];
-		float v1_z = v1.data()[2];
+		double v1_x = v1.data()[0];
+		double v1_y = v1.data()[1];
+		double v1_z = v1.data()[2];
 
-		float v2_x = v2.data()[0];
-		float v2_y = v2.data()[1];
-		float v2_z = v2.data()[2];
+		double v2_x = v2.data()[0];
+		double v2_y = v2.data()[1];
+		double v2_z = v2.data()[2];
 
 		return is_equal(v1_x, v2_x) && is_equal(v1_y, v2_y) && is_equal(v1_z, v2_z);
 	}
 
 	//// Convert radians to degrees
-	inline float rad2deg(float value) { return value * (180.0f / M_PI); }
+	inline double rad2deg(double value) { return value * (180.0f / M_PI); }
 
 	/// Convert degrees to radians
-	inline float deg2rad(float value) { return value * (M_PI / 180.0f); }
+	inline double deg2rad(double value) { return value * (M_PI / 180.0f); }
 
-	///Clamp a float to be between 0.0 and 1.0.
-    inline float clamp(float x){ return x<0.0f ? 0.0f : x>1.0f ? 1.0f : x; }
+	///Clamp a double to be between 0.0 and 1.0.
+    inline double clamp(double x){ return x<0.0f ? 0.0f : x>1.0f ? 1.0f : x; }
     
-    ///Cast a float between 0.0 and 1.0 to a char. A gamme value of 2.2 is used.
-    inline char to_char(float x){ return char(pow(clamp(x),1/2.2)*255+.5); }
+    ///Cast a double between 0.0 and 1.0 to a char. A gamme value of 2.2 is used.
+    inline char to_char(double x){ return char(pow(clamp(x),1/2.2)*255+.5); }
 }
